EPMRPP-104824 || Code Review fix - 3
