+++ app/src/components/main/dateRange/utils.js
@@ -15,6 +15,7 @@
  */
 
 import moment from 'moment';
+import { getMinutesFromTimestamp } from 'common/utils';
 import { DATE_FORMAT_DROPDOWN } from 'common/constants/timeDateFormat';
 
 export const parseFormattedDate = (formatted) => {
@@ -42,16 +43,16 @@ export const parseFormattedDate = (formatted) => {
   };
 };
 
-export const formatDisplayedValue = (displayedValue, lastRunDate, timeRangeValues) => {
-  if (!lastRunDate) {
+export const formatDisplayedValue = (displayedValue, value, timeRangeValues) => {
+  if (!value) {
     return displayedValue;
   }
 
-  if (timeRangeValues.includes(lastRunDate)) {
+  if (timeRangeValues.includes(value)) {
     return displayedValue;
   }
 
-  const { startDate, endDate } = parseFormattedDate(lastRunDate) || {};
+  const { startDate, endDate } = parseFormattedDate(value) || {};
 
   if (!startDate && !endDate) {
     return '';
@@ -62,4 +63,19 @@ export const formatDisplayedValue = (displayedValue, lastRunDate, timeRangeValue
   const formattedEndTimeRange = endDate && moment(new Date(endDate)).format(DATE_FORMAT_DROPDOWN);
 
   return `${formattedStartTimeRange || ''} — ${formattedEndTimeRange || ''}`;
+};
+
+export const formatDateRangeToMinutesString = (formValue) => {
+  if (typeof formValue === 'string') {
+    return formValue;
+  }
+
+  const { startDate, endDate } = formValue || {};
+  if (!startDate || !endDate) {
+    return '';
+  }
+
+  const utcString = moment().format('ZZ');
+
+  return `${getMinutesFromTimestamp(startDate)};${getMinutesFromTimestamp(endDate)};${utcString}`;
 };
