+++ app/src/components/main/filterButton/filterContent/filterContent.jsx
@@ -23,7 +23,6 @@ import { Button } from '@reportportal/ui-kit';
 import { useIntl } from 'react-intl';
 import PropTypes from 'prop-types';
 import { COMMON_LOCALE_KEYS } from 'common/constants/localization';
-import { getFormattedDate } from 'components/filterEntities/utils';
 import { FilterInput } from './filterInput/filterInput';
 import { FILTER_FORM } from '../constants';
 import { messages } from './messages';
