+++ app/src/pages/organization/organizationProjectsPage/projectsPageHeader/projectsFilter/projectsFilter.jsx
@@ -40,9 +40,9 @@ import {
   DateRangeFormField,
   formatDisplayedValue,
   parseFormattedDate,
+  formatDateRangeToMinutesString,
 } from 'components/main/dateRange';
 import classNames from 'classnames/bind';
-import { getFormattedDate } from 'components/filterEntities/utils';
 import { messages } from './messages';
 import styles from './projectsFilter.scss';
 
@@ -83,7 +83,7 @@ export const ProjectsFilter = ({
                 name={LAST_RUN_DATE_FILTER_NAME}
                 component={DateRangeFormField}
                 format={parseFormattedDate}
-                parse={getFormattedDate}
+                parse={formatDateRangeToMinutesString}
               />
             ),
           },
@@ -190,10 +190,6 @@ export const ProjectsFilter = ({
       return false;
     }
 
-    if (typeof lastRunDate === 'object' && (!lastRunDate.startDate || !lastRunDate?.endDate)) {
-      return true;
-    }
-
     let isApply = [LAUNCHES_FILTER_NAME, TEAMMATES_FILTER_NAME].every(
       (prop) => formValues[prop] === initialFilterState[prop],
     );
@@ -207,16 +203,6 @@ export const ProjectsFilter = ({
       isApply =
         formValues[TEAMMATES_FILTER_NAME_CONDITION] ===
           initialFilterState[TEAMMATES_FILTER_NAME_CONDITION] && isApply;
-    }
-
-    if (typeof lastRunDate === 'object') {
-      const initialDate = formatDisplayedValue(initialFilterState[LAST_RUN_DATE_FILTER_NAME]);
-      const currentDate = formatDisplayedValue(lastRunDate);
-      isApply = initialDate === currentDate && isApply;
-    } else {
-      isApply =
-        formValues[LAST_RUN_DATE_FILTER_NAME] === initialFilterState[LAST_RUN_DATE_FILTER_NAME] &&
-        isApply;
     }
 
     return isApply;
